:memo: Fixing the types in the error section and small improvements
These improvements include:
- changing "to determine if you got" to "to determine if we got", as to 
match the "We" from the beginning of the sentence, line 340
- splitting line 342
- switching "you probably want the union type version" to "you probably 
want the error version", line 342 -> 343
- appending "otherwise, the type union might fit better." to line 342 -> 
343
- changing "Be mindful that when you address one possible performance 
problem that you don't introduce a different one." to "Be mindful that, 
when you address one possible performance problem, you don't introduce a 
different one.", line 344.
This last change avoids the repetition of "that".